[feat] 정답체크 List형태로 Body받게 변경 + 리턴 타입 int 로 변경 (정답시 1/ 오답시 0) + 미사용

See merge request kofia/etest-sap-backend!57

+++ eTest/src/main/java/com/tmax/eTest/Contents/service/AnswerServices.java
@@ -20,7 +20,7 @@ public class AnswerServices {
 	ProblemChoiceRepository probChoiceRepo;
 	
 	
-	public Map<String, Object> getProblemSolution(long problemID) throws Exception{
+	public Map<String, Object> getProblemSolution(Integer problemID) throws Exception{
 		Map<String, Object> output = new HashMap<String, Object>();
 		Optional<Problem> problemOpt = problemRepo.findById(problemID);
 		
@@ -34,7 +34,7 @@ public class AnswerServices {
 		return output;
 	}
 	
-	public Map<String, Object> getSolutionMaterial(long problemID) throws Exception{
+	public Map<String, Object> getSolutionMaterial(Integer problemID) throws Exception{
 		Map<String, Object> output = new HashMap<String, Object>();
 		Optional<Problem> problemOpt = problemRepo.findById(problemID);
